v1/requirements.txt
selenium==4.18.1
pillow==10.2.0
numpy==1.24.4 
opencv-python===4.11.0